Fix half-step tonal values being rounded when stored in FujifilmRecipe
IntegerField silently rounded Decimal values like highlight=-1.5 to -2
and shadow=+0.5 to 0 via round(float(s)) in _parse_numeric.

- Migrate highlight, shadow, color, sharpness, high_iso_nr, clarity and
  monochromatic fields from IntegerField to DecimalField(max_digits=4,
  decimal_places=1) (migration 0019)
- Replace round(float(s)) with Decimal(s) in _parse_numeric so half-steps
  are preserved exactly
- Add detect_recipe_exif_mismatches management command to cross-reference
  FujifilmRecipe values against their images' FujifilmExif records; ran
  with --fix to correct 9 existing recipes whose highlight/shadow had been
  rounded
- Update _to_int → _to_decimal in verify_camera_recipes for consistency
- Add regression tests: TestHalfStepTonalValues in test_operations.py
- Document DecimalField requirement in tonal_adjustments_exif_mapping.md

+++ src/data/models.py
@@ -178,14 +178,14 @@ class FujifilmRecipe(models.Model):
     white_balance = models.CharField(max_length=100)
     white_balance_red = models.IntegerField()
     white_balance_blue = models.IntegerField()
-    highlight = models.IntegerField(null=True, blank=True)
-    shadow = models.IntegerField(null=True, blank=True)
-    color = models.IntegerField(null=True, blank=True)
-    sharpness = models.IntegerField(null=True, blank=True)
-    high_iso_nr = models.IntegerField(null=True, blank=True)
-    clarity = models.IntegerField(null=True, blank=True)
-    monochromatic_color_warm_cool = models.IntegerField(null=True, blank=True)
-    monochromatic_color_magenta_green = models.IntegerField(null=True, blank=True)
+    highlight = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
+    shadow = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
+    color = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
+    sharpness = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
+    high_iso_nr = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
+    clarity = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
+    monochromatic_color_warm_cool = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
+    monochromatic_color_magenta_green = models.DecimalField(max_digits=4, decimal_places=1, null=True, blank=True)
 
     # Factories
